sparkplug-payload from require to import
so we can benefit from the typings supplied with
the sparkplug-payload library
move the filter function into a file of its own,
so the main function remains legible
