+++ src/sp_pub/sp_pub.ts
@@ -1,12 +1,12 @@
-import { connect, IClientOptions } from "async-mqtt";
-import { args } from "./args";
+import {connect, IClientOptions} from "async-mqtt";
+import {args} from "./args";
 import fs from "fs";
 
-const sparkplug = require('sparkplug-payload');
+import * as sparkplug from 'sparkplug-payload';
 const sparkplugbpayload = sparkplug.get("spBv1.0");
-const encodePayload = sparkplugbpayload.encodePayload;
+const encodePayload = sparkplugbpayload!.encodePayload;
 
-const { host, port, topic, message, cafile, key, cert, insecure, id } = args;
+const {host, port, topic, message, cafile, key, cert, insecure, id} = args;
 
 
 const mqttClientOptions: IClientOptions = {
