+++ src/sp_sub/sp_sub.ts
@@ -1,15 +1,18 @@
-import * as _ from "lodash";
-import { connect, IClientOptions } from "async-mqtt";
-import { Packet } from "mqtt-packet";
-import pako from "pako"; import { args } from "./args";
+import _ from "lodash";
+import {connect, IClientOptions} from "async-mqtt";
+import {Packet} from "mqtt-packet";
+import pako from "pako";
+import {args} from "./args";
 import fs from "fs";
 import Long from "long";
 
-const sparkplug = require('sparkplug-payload');
+import * as sparkplug from 'sparkplug-payload';
 const sparkplugbpayload = sparkplug.get("spBv1.0");
-const decodePayload = sparkplugbpayload.decodePayload;
+const decodePayload = sparkplugbpayload!.decodePayload;
 
-const { host, port, topic, gunzip, pretty, json, verbose, cafile, key, cert, insecure, id, includeMetric, excludeMetric, showTimestamp } = args;
+import {filter} from "./filter";
+
+const {host, port, topic, gunzip, pretty, json, verbose, cafile, key, cert, insecure, id, showTimestamp} = args;
 
 
 const mqttClientOptions: IClientOptions = {
@@ -50,34 +53,17 @@ const onMessage = async (topic: string, payload: Buffer, msg: Packet) => {
                 decoded = decodePayload(body);
             }
 
-            const filteredMetrics = [];
-            for (const metric of decoded.metrics) {
-                if (includeMetric !== undefined) {
-                    if (metric.name.match(includeMetric)) {
-                        filteredMetrics.push(metric);
-                    }
-                } else if (excludeMetric !== undefined) {
-                    if (!metric.name.match(excludeMetric)) {
-                        filteredMetrics.push(metric);
-                    }
-                } else {
-                    filteredMetrics.push(metric);
-                }
-            }
-            if (filteredMetrics.length === 0) {
-                return;
-            }
-            decoded.metrics = filteredMetrics;
+            decoded.metrics = filter(decoded.metrics);
 
             if (verbose) {
                 console.log();
-                console.log(`${showTimestamp ? new Date().toISOString() + " ": ""}${topic}`);
+                console.log(`${showTimestamp ? new Date().toISOString() + " " : ""}${topic}`);
             }
 
             if (json) {
                 console.log(JSON.stringify(decoded, (key, value) => (typeof value === "bigint") || (Long.isLong(value)) ? parseFloat(value.toString()) : value, pretty ? 2 : undefined));
             } else {
-                console.dir(decoded, { breakLength: Infinity, maxStringLength: Infinity, maxArrayLength: Infinity, compact: !pretty, depth: Infinity });
+                console.dir(decoded, {breakLength: Infinity, maxStringLength: Infinity, maxArrayLength: Infinity, compact: !pretty, depth: Infinity});
             }
         } catch (e: any) {
             console.error(e.message);
